Refactored AI directories
