+++ ai/langchain/langchain_factory.py
@@ -1,5 +1,5 @@
 from ai.ai_factory import AIFactory
-from ai.langchain import LangChainAI
+from ai.langchain.langchain import LangChainAI
 
 class LangChainAIFactory(AIFactory):
     def create(self, tools, params):
